update pseudogradients when generating x
Finished. See notes for details.
For each added agent, remove calls from lateCalls because they can be
picked up.
In simulation, keep track of the number of calls picked up by last, 2nd
to last, and 3rd to last agents in each group. Update the lastCalls list
according to how many agents are left.
This might be oversimplifying lastCalls - I wish to move lastCalls to
lateCalls when the agent is gone, since the call needs to be queued.